iq-cli-derive: Refactor into separate modules
Refactors the monolithic `lib.rs` originally sourced from `gumdrop` into
separate modules, in hopes of making it a little more maintainable.
