+++ iq-cli-derive/src/lib.rs
@@ -73,13 +73,25 @@ extern crate syn;
 use std::iter::repeat;
 
 use proc_macro::TokenStream;
-use proc_macro2::TokenStream as TokenStream2;
 
 use quote::ToTokens;
 use syn::{
-    parse_str, AttrStyle, Attribute, Data, DataEnum, DataStruct, DeriveInput, Fields,
-    GenericArgument, Ident, Lit, Meta, NestedMeta, Path, PathArguments, Type,
+    Data, DataEnum, DataStruct, DeriveInput, Fields, GenericArgument, Lit, PathArguments, Type,
 };
+
+mod action;
+mod opts;
+mod parse;
+
+use action::{Action, FreeAction};
+use opts::{
+    attr::AttrOpts,
+    cmd::{Cmd, CmdOpts},
+    default::DefaultOpts,
+    free::FreeOpt,
+    opt::Opt,
+};
+use parse::ParseFn;
 
 /// Derive the `Options` trait
 #[proc_macro_derive(Options, attributes(options))]
@@ -624,717 +636,7 @@ fn derive_options_struct(ast: &DeriveInput, fields: &Fields) -> TokenStream {
     expr.to_string().parse().expect("parse quote!")
 }
 
-enum Action {
-    /// Increase count
-    Count,
-    /// Push an argument to a `Vec<T>` field
-    Push(ParseMethod),
-    /// Set field
-    SetField(ParseMethod),
-    /// Set `Option<T>` field
-    SetOption(ParseMethod),
-    /// Set field to `true`
-    Switch,
-}
-
-#[derive(Default)]
-struct AttrOpts {
-    long: Option<String>,
-    short: Option<char>,
-    free: bool,
-    count: bool,
-    help_flag: bool,
-    no_help_flag: bool,
-    no_short: bool,
-    no_long: bool,
-    required: bool,
-    not_required: bool,
-    help: Option<String>,
-    meta: Option<String>,
-    parse: Option<ParseFn>,
-    default: Option<String>,
-
-    command: bool,
-}
-
-struct Cmd<'a> {
-    name: String,
-    help: Option<String>,
-    variant_name: &'a Ident,
-    ty: &'a Type,
-}
-
-#[derive(Default)]
-struct CmdOpts {
-    name: Option<String>,
-    help: Option<String>,
-}
-
-#[derive(Default)]
-struct DefaultOpts {
-    no_help_flag: bool,
-    no_long: bool,
-    no_short: bool,
-    required: bool,
-}
-
-#[derive(Copy, Clone, Eq, PartialEq)]
-enum FreeAction {
-    Push,
-    SetField,
-    SetOption,
-}
-
-struct FreeOpt<'a> {
-    field: &'a Ident,
-    action: FreeAction,
-    parse: ParseFn,
-    required: bool,
-    help: Option<String>,
-}
-
-struct Opt<'a> {
-    field: &'a Ident,
-    action: Action,
-    long: Option<String>,
-    short: Option<char>,
-    no_short: bool,
-    required: bool,
-    help: Option<String>,
-    meta: Option<String>,
-    default: Option<String>,
-}
-
-enum ParseFn {
-    Default,
-    FromStr(Option<Path>),
-    TryFromStr(Path),
-}
-
-struct ParseMethod {
-    parse_fn: ParseFn,
-    tuple_len: Option<usize>,
-}
-
-impl Action {
-    fn infer(ty: &Type, parse: Option<ParseFn>) -> Action {
-        match *ty {
-            Type::Path(ref path) => {
-                let path = path.path.segments.last().unwrap().into_value();
-                let param = first_ty_param(ty);
-
-                match &path.ident.to_string()[..] {
-                    "bool" if parse.is_none() => Action::Switch,
-                    "Vec" if param.is_some() => {
-                        let tuple_len = tuple_len(param.unwrap());
-
-                        Action::Push(ParseMethod {
-                            parse_fn: parse.unwrap_or_default(),
-                            tuple_len,
-                        })
-                    }
-                    "Option" if param.is_some() => {
-                        let tuple_len = tuple_len(param.unwrap());
-
-                        Action::SetOption(ParseMethod {
-                            parse_fn: parse.unwrap_or_default(),
-                            tuple_len,
-                        })
-                    }
-                    _ => Action::SetField(ParseMethod {
-                        parse_fn: parse.unwrap_or_default(),
-                        tuple_len: tuple_len(ty),
-                    }),
-                }
-            }
-            _ => {
-                let tuple_len = tuple_len(ty);
-
-                Action::SetField(ParseMethod {
-                    parse_fn: parse.unwrap_or_default(),
-                    tuple_len,
-                })
-            }
-        }
-    }
-
-    fn takes_arg(&self) -> bool {
-        use self::Action::*;
-
-        match *self {
-            Push(ref meth) | SetField(ref meth) | SetOption(ref meth) => meth.takes_arg(),
-            _ => false,
-        }
-    }
-
-    fn tuple_len(&self) -> Option<usize> {
-        use self::Action::*;
-
-        match *self {
-            Push(ref meth) | SetField(ref meth) | SetOption(ref meth) => meth.tuple_len,
-            _ => None,
-        }
-    }
-}
-
-impl AttrOpts {
-    #[allow(unknown_lints, cyclomatic_complexity)]
-    fn check(&self) {
-        if self.command {
-            if self.free {
-                panic!("`command` and `free` are mutually exclusive");
-            }
-            if self.default.is_some() {
-                panic!("`command` and `default` are mutually exclusive");
-            }
-            if self.long.is_some() {
-                panic!("`command` and `long` are mutually exclusive");
-            }
-            if self.short.is_some() {
-                panic!("`command` and `short` are mutually exclusive");
-            }
-            if self.count {
-                panic!("`command` and `count` are mutually exclusive");
-            }
-            if self.help_flag {
-                panic!("`command` and `help_flag` are mutually exclusive");
-            }
-            if self.no_help_flag {
-                panic!("`command` and `no_help_flag` are mutually exclusive");
-            }
-            if self.no_short {
-                panic!("`command` and `no_short` are mutually exclusive");
-            }
-            if self.no_long {
-                panic!("`command` and `no_long` are mutually exclusive");
-            }
-            if self.help.is_some() {
-                panic!("`command` and `help` are mutually exclusive");
-            }
-            if self.meta.is_some() {
-                panic!("`command` and `meta` are mutually exclusive");
-            }
-        }
-
-        if self.free {
-            if self.default.is_some() {
-                panic!("`free` and `default` are mutually exclusive");
-            }
-            if self.long.is_some() {
-                panic!("`free` and `long` are mutually exclusive");
-            }
-            if self.short.is_some() {
-                panic!("`free` and `short` are mutually exclusive");
-            }
-            if self.count {
-                panic!("`free` and `count` are mutually exclusive");
-            }
-            if self.help_flag {
-                panic!("`free` and `help_flag` are mutually exclusive");
-            }
-            if self.no_help_flag {
-                panic!("`free` and `no_help_flag` are mutually exclusive");
-            }
-            if self.no_short {
-                panic!("`free` and `no_short` are mutually exclusive");
-            }
-            if self.no_long {
-                panic!("`free` and `no_long` are mutually exclusive");
-            }
-            if self.meta.is_some() {
-                panic!("`free` and `meta` are mutually exclusive");
-            }
-        }
-
-        if self.help_flag && self.no_help_flag {
-            panic!("`help_flag` and `no_help_flag` are mutually exclusive");
-        }
-
-        if self.no_short && self.short.is_some() {
-            panic!("`no_short` and `short` are mutually exclusive");
-        }
-
-        if self.no_long && self.long.is_some() {
-            panic!("`no_long` and `long` are mutually exclusive");
-        }
-
-        if self.required && self.not_required {
-            panic!("`required` and `not_required` are mutually exclusive");
-        }
-
-        if self.parse.is_some() && self.count {
-            panic!("`count` and `parse` are mutually exclusive");
-        }
-    }
-
-    fn parse(attrs: &[Attribute]) -> AttrOpts {
-        let mut opts = AttrOpts::default();
-
-        for attr in attrs {
-            if is_outer(attr.style) && path_eq(&attr.path, "options") {
-                let meta = attr
-                    .interpret_meta()
-                    .unwrap_or_else(|| panic!("invalid attribute: {}", tokens_str(attr)));
-
-                match meta {
-                    Meta::Word(_) => panic!("#[options] is not a valid attribute"),
-                    Meta::NameValue(..) => panic!("#[options = ...] is not a valid attribute"),
-                    Meta::List(ref items) => {
-                        for item in &items.nested {
-                            opts.parse_item(item);
-                        }
-                    }
-                }
-            }
-        }
-
-        opts.check();
-
-        opts
-    }
-
-    fn parse_item(&mut self, item: &NestedMeta) {
-        match *item {
-            NestedMeta::Literal(_) => panic!("unexpected meta item `{}`", tokens_str(item)),
-            NestedMeta::Meta(ref item) => match *item {
-                Meta::Word(ref w) => match &w.to_string()[..] {
-                    "free" => self.free = true,
-                    "command" => self.command = true,
-                    "count" => self.count = true,
-                    "help_flag" => self.help_flag = true,
-                    "no_help_flag" => self.no_help_flag = true,
-                    "no_short" => self.no_short = true,
-                    "no_long" => self.no_long = true,
-                    "required" => self.required = true,
-                    "not_required" => self.not_required = true,
-                    _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-                },
-                Meta::List(ref list) => match &list.ident.to_string()[..] {
-                    "parse" => {
-                        if list.nested.len() != 1 {
-                            panic!("unexpected meta item `{}`", tokens_str(item));
-                        }
-
-                        self.parse = Some(ParseFn::parse(&list.nested[0]));
-                    }
-                    _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-                },
-                Meta::NameValue(ref nv) => match &nv.ident.to_string()[..] {
-                    "default" => self.default = Some(lit_str(&nv.lit)),
-                    "long" => self.long = Some(lit_str(&nv.lit)),
-                    "short" => self.short = Some(lit_char(&nv.lit)),
-                    "help" => self.help = Some(lit_str(&nv.lit)),
-                    "meta" => self.meta = Some(lit_str(&nv.lit)),
-                    _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-                },
-            },
-        }
-    }
-
-    fn set_defaults(&mut self, defaults: &DefaultOpts) {
-        if !self.help_flag && defaults.no_help_flag {
-            self.no_help_flag = true;
-        }
-        if self.short.is_none() && defaults.no_short {
-            self.no_short = true;
-        }
-        if self.long.is_none() && defaults.no_long {
-            self.no_long = true;
-        }
-
-        if self.not_required {
-            self.required = false;
-        } else if defaults.required {
-            self.required = true;
-        }
-    }
-}
-
-impl CmdOpts {
-    fn parse(attrs: &[Attribute]) -> CmdOpts {
-        let mut opts = CmdOpts::default();
-
-        for attr in attrs {
-            if is_outer(attr.style) && path_eq(&attr.path, "options") {
-                let meta = attr
-                    .interpret_meta()
-                    .unwrap_or_else(|| panic!("invalid attribute: {}", tokens_str(attr)));
-
-                match meta {
-                    Meta::Word(_) => panic!("#[options] is not a valid attribute"),
-                    Meta::NameValue(..) => panic!("#[options = ...] is not a valid attribute"),
-                    Meta::List(ref items) => {
-                        for item in &items.nested {
-                            match *item {
-                                NestedMeta::Literal(_) => {
-                                    panic!("unexpected meta item `{}`", tokens_str(item))
-                                }
-                                NestedMeta::Meta(ref item) => match *item {
-                                    Meta::Word(_) | Meta::List(..) => {
-                                        panic!("unexpected meta item `{}`", tokens_str(item))
-                                    }
-                                    Meta::NameValue(ref nv) => match &nv.ident.to_string()[..] {
-                                        "name" => opts.name = Some(lit_str(&nv.lit)),
-                                        "help" => opts.help = Some(lit_str(&nv.lit)),
-                                        _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-                                    },
-                                },
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        opts
-    }
-}
-
-impl DefaultOpts {
-    fn parse(attrs: &[Attribute]) -> DefaultOpts {
-        let mut opts = DefaultOpts::default();
-
-        for attr in attrs {
-            if is_outer(attr.style) && path_eq(&attr.path, "options") {
-                let meta = attr
-                    .interpret_meta()
-                    .unwrap_or_else(|| panic!("invalid attribute: {}", tokens_str(attr)));
-
-                match meta {
-                    Meta::Word(_) => panic!("#[options] is not a valid attribute"),
-                    Meta::NameValue(..) => panic!("#[options = ...] is not a valid attribute"),
-                    Meta::List(ref items) => {
-                        for item in &items.nested {
-                            match *item {
-                                NestedMeta::Literal(_) => {
-                                    panic!("unexpected meta item `{}`", tokens_str(item))
-                                }
-                                NestedMeta::Meta(ref item) => match *item {
-                                    Meta::Word(ref w) => match &w.to_string()[..] {
-                                        "no_help_flag" => opts.no_help_flag = true,
-                                        "no_short" => opts.no_short = true,
-                                        "no_long" => opts.no_long = true,
-                                        "required" => opts.required = true,
-                                        _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-                                    },
-                                    Meta::List(..) | Meta::NameValue(..) => {
-                                        panic!("unexpected meta item `{}`", tokens_str(item))
-                                    }
-                                },
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        opts
-    }
-}
-
-impl FreeAction {
-    fn infer(ty: &Type) -> FreeAction {
-        match *ty {
-            Type::Path(ref path) => {
-                let path = path.path.segments.last().unwrap().into_value();
-
-                match &path.ident.to_string()[..] {
-                    "Option" => FreeAction::SetOption,
-                    "Vec" => FreeAction::Push,
-                    _ => FreeAction::SetField,
-                }
-            }
-            _ => FreeAction::SetField,
-        }
-    }
-}
-
-impl<'a> FreeOpt<'a> {
-    fn mark_used(&self) -> TokenStream2 {
-        if self.required {
-            let field = self.field;
-            quote!{ _used.#field = true; }
-        } else {
-            quote!{}
-        }
-    }
-
-    fn width(&self) -> usize {
-        2 + self.field.to_string().len() + 2 // name + spaces before and after
-    }
-}
-
-impl<'a> Opt<'a> {
-    fn display_form(&self) -> String {
-        if let Some(ref long) = self.long {
-            format!("--{}", long)
-        } else {
-            format!("-{}", self.short.unwrap())
-        }
-    }
-
-    fn mark_used(&self) -> TokenStream2 {
-        if self.required {
-            let field = self.field;
-            quote!{ _used.#field = true; }
-        } else {
-            quote!{}
-        }
-    }
-
-    fn width(&self) -> usize {
-        let short = self.short.map_or(0, |_| 1 + 1); // '-' + char
-        let long = self.long.as_ref().map_or(0, |s| s.len() + 2); // "--" + str
-        let sep = if short == 0 || long == 0 { 0 } else { 2 }; // ", "
-        let meta = self.meta.as_ref().map_or(0, |s| s.len() + 1); // ' ' + meta
-
-        2 + short + long + sep + meta + 2 // total + spaces before and after
-    }
-
-    fn make_action(&self) -> TokenStream2 {
-        use self::Action::*;
-
-        let field = self.field;
-        let mark_used = self.mark_used();
-
-        let action = match self.action {
-            Count => quote!{
-                _result.#field += 1;
-            },
-            Push(ref meth) => {
-                let act = meth.make_action_type();
-
-                quote!{
-                    _result.#field.push(#act);
-                }
-            }
-            SetField(ref meth) => {
-                let act = meth.make_action_type();
-
-                quote!{
-                    _result.#field = #act;
-                }
-            }
-            SetOption(ref meth) => {
-                let act = meth.make_action_type();
-
-                quote!{
-                    _result.#field = ::std::option::Option::Some(#act);
-                }
-            }
-            Switch => quote!{
-                _result.#field = true;
-            },
-        };
-
-        quote!{
-            #mark_used
-            #action
-        }
-    }
-
-    fn make_action_arg(&self) -> TokenStream2 {
-        use self::Action::*;
-
-        let field = self.field;
-        let mark_used = self.mark_used();
-
-        let action = match self.action {
-            Push(ref meth) => {
-                let act = meth.make_action_type_arg();
-
-                quote!{
-                    _result.#field.push(#act);
-                }
-            }
-            SetField(ref meth) => {
-                let act = meth.make_action_type_arg();
-
-                quote!{
-                    _result.#field = #act;
-                }
-            }
-            SetOption(ref meth) => {
-                let act = meth.make_action_type_arg();
-
-                quote!{
-                    _result.#field = ::std::option::Option::Some(#act);
-                }
-            }
-            _ => unreachable!(),
-        };
-
-        quote!{
-            #mark_used
-            #action
-        }
-    }
-
-    fn usage(&self, col_width: usize) -> String {
-        let mut res = String::from("  ");
-
-        if let Some(short) = self.short {
-            res.push('-');
-            res.push(short);
-        }
-
-        if self.short.is_some() && self.long.is_some() {
-            res.push_str(", ");
-        }
-
-        if let Some(ref long) = self.long {
-            res.push_str("--");
-            res.push_str(long);
-        }
-
-        if let Some(ref meta) = self.meta {
-            res.push(' ');
-            res.push_str(meta);
-        }
-
-        if self.help.is_some() || self.default.is_some() {
-            if res.len() < col_width {
-                let n = col_width - res.len();
-                res.extend(repeat(' ').take(n));
-            } else {
-                res.push('\n');
-                res.extend(repeat(' ').take(col_width));
-            }
-        }
-
-        if let Some(ref help) = self.help {
-            res.push_str(help);
-        }
-
-        if let Some(ref default) = self.default {
-            res.push_str(" (default: ");
-            res.push_str(default);
-            res.push_str(")");
-        }
-
-        res
-    }
-}
-
-impl ParseFn {
-    fn parse(item: &NestedMeta) -> ParseFn {
-        match *item {
-            NestedMeta::Meta(Meta::Word(ref ident)) => match &ident.to_string()[..] {
-                "from_str" => ParseFn::FromStr(None),
-                "try_from_str" => ParseFn::Default,
-                _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-            },
-            NestedMeta::Meta(Meta::NameValue(ref nv)) => match &nv.ident.to_string()[..] {
-                "from_str" => {
-                    let path = parse_str(&lit_str(&nv.lit)).unwrap();
-                    ParseFn::FromStr(Some(path))
-                }
-                "try_from_str" => {
-                    let path = parse_str(&lit_str(&nv.lit)).unwrap();
-                    ParseFn::TryFromStr(path)
-                }
-                _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-            },
-            _ => panic!("unexpected meta item `{}`", tokens_str(item)),
-        }
-    }
-
-    fn make_parse_action(&self) -> TokenStream2 {
-        match *self {
-            ParseFn::Default => quote!{
-                ::std::str::FromStr::from_str(_arg)
-                    .map_err(|e| ::iq_cli::options::Error::failed_parse(_opt,
-                        ::std::string::ToString::to_string(&e)))?
-            },
-            ParseFn::FromStr(None) => quote!{
-                ::std::convert::From::from(_arg)
-            },
-            ParseFn::FromStr(Some(ref fun)) => quote!{
-                #fun(_arg)
-            },
-            ParseFn::TryFromStr(ref fun) => quote!{
-                #fun(_arg)
-                    .map_err(|e| ::iq_cli::options::Error::failed_parse(_opt,
-                        ::std::string::ToString::to_string(&e)))?
-            },
-        }
-    }
-
-    fn make_parse_default_action(&self, ident: &Ident, expr: &str) -> TokenStream2 {
-        match *self {
-            ParseFn::Default => quote!{
-                ::std::str::FromStr::from_str(#expr)
-                    .map_err(|e| ::iq_cli::options::Error::failed_parse_default(
-                        stringify!(#ident), #expr,
-                        ::std::string::ToString::to_string(&e)))?
-            },
-            ParseFn::FromStr(None) => quote!{
-                ::std::convert::From::from(#expr)
-            },
-            ParseFn::FromStr(Some(ref fun)) => quote!{
-                #fun(#expr)
-            },
-            ParseFn::TryFromStr(ref fun) => quote!{
-                #fun(#expr)
-                    .map_err(|e| ::iq_cli::options::Error::failed_parse_default(
-                        stringify!(#ident), #expr,
-                        ::std::string::ToString::to_string(&e)))?
-            },
-        }
-    }
-}
-
-impl Default for ParseFn {
-    fn default() -> ParseFn {
-        ParseFn::Default
-    }
-}
-
-impl ParseMethod {
-    fn make_action_type(&self) -> TokenStream2 {
-        let parse = self.parse_fn.make_parse_action();
-
-        match self.tuple_len {
-            None => quote!{ {
-                let _arg = _parser.next_arg()
-                    .ok_or_else(|| ::iq_cli::options::Error::missing_argument(_opt))?;
-
-                #parse
-            } },
-            Some(n) => {
-                let num = 0..n;
-                let n = repeat(n);
-                let parse = repeat(parse);
-
-                quote!{
-                    ( #( {
-                        let _found = #num;
-                        let _arg = _parser.next_arg()
-                            .ok_or_else(|| ::iq_cli::options::Error::insufficient_arguments(
-                                _opt, #n, _found))?;
-
-                        #parse
-                    } , )* )
-                }
-            }
-        }
-    }
-
-    fn make_action_type_arg(&self) -> TokenStream2 {
-        match self.tuple_len {
-            None => self.parse_fn.make_parse_action(),
-            Some(_) => unreachable!(),
-        }
-    }
-    fn takes_arg(&self) -> bool {
-        match self.tuple_len {
-            Some(0) => false,
-            _ => true,
-        }
-    }
-}
-
-fn first_ty_param(ty: &Type) -> Option<&Type> {
+pub(crate) fn first_ty_param(ty: &Type) -> Option<&Type> {
     match *ty {
         Type::Path(ref path) => {
             let path = path.path.segments.last().unwrap().into_value();
@@ -1358,59 +660,15 @@ fn first_ty_param(ty: &Type) -> Option<&Type> {
     }
 }
 
-fn is_outer(style: AttrStyle) -> bool {
-    match style {
-        AttrStyle::Outer => true,
-        _ => false,
-    }
-}
-
-fn lit_str(lit: &Lit) -> String {
+pub(crate) fn lit_str(lit: &Lit) -> String {
     match *lit {
         Lit::Str(ref s) => s.value(),
         _ => panic!("unexpected literal `{}`", tokens_str(lit)),
     }
 }
 
-fn lit_char(lit: &Lit) -> char {
-    match *lit {
-        // Character literals in attributes are not necessarily allowed
-        Lit::Str(ref s) => {
-            let s = s.value();
-            let mut chars = s.chars();
-
-            let res = chars.next().expect("expected one-char string literal");
-            if chars.next().is_some() {
-                panic!("expected one-char string literal");
-            }
-
-            res
-        }
-        Lit::Char(ref ch) => ch.value(),
-        _ => panic!("unexpected literal `{}`", tokens_str(lit)),
-    }
-}
-
-fn path_eq(path: &Path, s: &str) -> bool {
-    path.segments.len() == 1 && {
-        let seg = path.segments.first().unwrap().into_value();
-
-        match seg.arguments {
-            PathArguments::None => seg.ident == s,
-            _ => false,
-        }
-    }
-}
-
-fn tokens_str<T: ToTokens>(t: &T) -> String {
+pub(crate) fn tokens_str<T: ToTokens>(t: &T) -> String {
     t.into_token_stream().to_string()
-}
-
-fn tuple_len(ty: &Type) -> Option<usize> {
-    match *ty {
-        Type::Tuple(ref tup) => Some(tup.elems.len()),
-        _ => None,
-    }
 }
 
 fn make_command_name(name: &str) -> String {
